feat(ui): move Scheduling to its own page and nav item
